requirements.txt
# PyTorch with CUDA 12.1 support
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision

# Core dependencies
numpy>=1.20.0
pyyaml>=5.4.0
tqdm>=4.60.0
pytest>=7.0.0
